remove unused kernel
